requirements.txt
Flask==2.0.2
Werkzeug==2.0.3
gunicorn
discord.py
requests
python-dotenv
mojito2
yfinance
finance-datareader
tradingview_ta
investpy
bs4
datetime
pyyaml
typing
matplotlib
asyncio
plotly
tabulate
yahoo_earnings_calendar
yahoo_fin 
requests_html
mplfinance
flask_cors
pandas==2.2.1
pandas_ta==0.3.14b0
numpy==1.26.4
requests_html-0.10.0-py3-none-any.whl
mplchart-0.0.8-py3-none-any.whl
certifi
gunicorn
celery
redis
